+++ src/org/epics/ioc/channelAccess/MonitorOnAbsoluteChangeFactory.java
@@ -42,7 +42,7 @@ public class MonitorOnAbsoluteChangeFactory {
                 ChannelMonitorRequester channelMonitorRequester,
                 PVStructure pvOption,
                 PVCopy pvCopy,
-                byte queueSize,
+                int queueSize,
                 Executor executor)
         {
             PVStructure pvStructure = pvCopy.createPVStructure();
@@ -82,7 +82,7 @@ public class MonitorOnAbsoluteChangeFactory {
                 Channel channel,
                 ChannelMonitorRequester channelMonitorRequester,
                 PVCopy pvCopy,
-                byte queueSize,
+                int queueSize,
                 Executor executor,
                 double deadband,
                 PVScalar valuePVField)
